requirements.txt
apache-beam==2.27.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.6.0
avro-python3==1.9.2.1
blinker==1.4
cachetools==4.0.0
certifi==2018.8.24
cffi==1.14.4
chardet==3.0.4
Click==7.0
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==2.6.1
cycler==0.10.0
dbus-python==1.2.12
dill==0.3.1.1
distro==1.3.0
distro-info===0.21ubuntu4
docopt==0.6.2
entrypoints==0.3
fastavro==1.2.3
fasteners==0.16
future==0.18.2
google-api-core==1.25.0
google-apitools==0.5.31
google-auth==1.24.0
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.6.1
google-cloud-build==2.0.0
google-cloud-core==1.5.0
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-storage==1.35.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-compute-engine==20190801.0
google-crc32c==1.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.1
grpcio-gcp==0.2.2
hdfs==2.5.8
httplib2==0.11.3
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.5.0
incremental==16.10.1
install==1.3.4
Jinja2==2.10
joblib==0.14.1
jsonpatch==1.22
jsonpointer==1.10
jsonschema==2.6.0
keyring==18.0.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.16
MarkupSafe==1.1.0
matplotlib==3.2.0
mecab-python3==1.0.3
mock==2.0.0
more-itertools==8.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.18.1
oauth2client==4.1.3
oauthlib==2.1.0
packaging==20.3
pbr==5.4.4
Pillow==7.0.0
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
py==1.8.1
pyarrow==2.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pydot==1.4.1
PyGObject==3.34.0
PyJWT==1.7.0
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyserial==3.4
pytest==5.4.1
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.36
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
requests==2.25.1
requests-unixsocket==0.1.5
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.7
systemd-python==234
testresources==2.0.1
Twisted==18.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ubuntu-advantage-tools==19.5
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
wadllib==1.3.3
wcwidth==0.1.8
wordcloud==1.6.0
zipp==3.1.0
zope.interface==4.3.2
